requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
gunicorn==23.0.0

transformers==4.46.3
huggingface_hub==0.26.2
tokenizers==0.20.3
safetensors==0.4.5

# CPU PyTorch (install via extra index in Dockerfile/local instructions)
# torch is intentionally not pinned here to avoid pip index differences
# but we pin it in Dockerfile. For local: see README.

prometheus-client==0.21.0
python-json-logger==2.0.7

opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-instrumentation-logging==0.49b2

pydantic==2.9.2
httpx==0.27.2

pytest==8.3.3
pytest-cov==5.0.0
